"Uploaded alarm clock project"
This commit contains the complete Python Alarm Clock project developed by Shourya Singh, VIT Bhopal,  for the Python Essentials course. The project allows the user to set an alarm using hours and minutes, and plays an audio alert using the pyttsx3 text-to-speech module when the alarm time is reached. This upload includes the main Python script and all required project files for Vidyarthi submission.